Charging modes, Vfloat Ramping, Testing Mode
None of the functions below are run in code yet, but all have been tested
using the new testing mode

Charging modes:
powerManager class now has methods to control charging mode (suspend,
resume, disable, enable), important to note that these overwrite vfloat's value, setting it to 3.8.

Vfloat ramping:
function that accepts a float and ramps vfloat depending on the value of
the input.

Testing mode:
The only agent-side change in this commit is one that handles the output
of any test we run. To run a test, replace unitTest() at the bottom with
your test, and set runTest to true at the top of sensor.device.

Helper functions:

Power manager has new helper methods for converting register readings into
either a string (example: "FC") or and array of single character strings
(example: ["F","C"]), though the latter should be changed to return an
array of chars.

toHexStr() is a function that accepts two strings as it's inputs and
outputs a string that can be passed to registers as a valid write value.