Tempcontrol added, Fixed register writing
Added Tempcontrol which disables the charger or issues warnings if the
temperature is above/below/near unacceptable temperatures.

Fixed power mode and changevfloat methods to no longer overwrite the half
of the register that does not concern them

added PMRegB and PMRegC to nv.data to keep track of any power manager mode
changes.

Added tempcontrol and changevfloat to main function so that they actually
run and do things now.